Add 'Pipe RGB to ImageMagick' macro
Pipe the script in the current tab to ImageMagick as RGB24 or RGB48.  The
frames can be sent in batches; various splitting options are provided.  The
macro runs on its own thread and doesn't use the clip utilised for the video
preview, so AvsPmod can still be used at the same time.

Originated from this discussion:
<http://forum.doom9.org/showthread.php?p=1598231#post1598231>